Move SDK init methods
